[#230] :recycle: refactor(types): input type 분리, 유틸리티 타입 활용
