--- client/src/components/diary/DiaryInput.ts
@@ -1,5 +1,0 @@
-export interface DiaryInput {
-  readonly title: string;
-  readonly content: string;
-}
-// TODO: 다이어리 폼, 다이어리 수정 폼에서 쓰임
