+++ client/src/components/new-book/ContainerTitle.tsx
@@ -3,7 +3,6 @@ import React from "react";
 interface TitleProps {
   children: string;
 }
-// TODO: input type에서 가져오기?
 const ContainerTitle = ({ children }: TitleProps) => {
   return (
     <div className="font-sans font-bold w-[75%] mx-auto mt-[2vh] text-[2vh]">{children}</div>
