+++ client/src/components/new-diary/NewDiaryForm.tsx
@@ -2,13 +2,13 @@ import React from "react";
 import { useRecoilValue } from "recoil";
 import { useForm } from "react-hook-form";
 
+import { DiaryInput } from "src/@types/input";
 import DiaryImgUpload from "./DiaryImgUpload";
 import { diaryImgAtom } from "src/recoil/new-diary/atom";
 import { Img } from "src/recoil/new-diary/atom";
 import { API_URL } from "src/constants/API_URL";
 import { useNavigate, useParams } from "react-router-dom";
 import { DiaryForm, TitleInput, inputStyle, ContentInput } from "../diary/DiaryFormElem";
-import { DiaryInput } from "../diary/DiaryInput";
 import { formApi } from "src/utils/axiosApi/formApi";
 
 const createDiaryForm = (diaryImg: Img, bookId: string, { title, content }: DiaryInput) => {
